Final Fix Implemented H2 DATABASE changed int for long

+++ src/main/java/com/ecomarket/ecomarket/service/ClientesService.java
@@ -25,10 +25,10 @@ public class ClientesService {
     }
 
     public Optional<Clientes> buscarPorId(int id) {
-        return clientesRepository.findById(id);
+        return clientesRepository.findById((long) id);
     }
 
     public void eliminarCliente(int id) {
-        clientesRepository.deleteById(id);
+        clientesRepository.deleteById((long) id);
     }
 }
